`Added middleware and updated routes for authentication and authorization`

+++ routes/web.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Http\Controllers\Auth\LoginController;
 use App\Http\Controllers\Auth\LogOutController;
 use App\Http\Controllers\Auth\RegisterController;
 use Illuminate\Support\Facades\Route;
@@ -36,15 +37,17 @@ Route::get('/Home/Contact', function () {
 
 Route::get("/Admin/Profile", function(){
     return view("Admin.Profile");
-});
+})->middleware("auth");
 
 Route::get("/Admin/Logs", function(){
     return view("Admin.AdminLogs");
-});
+})->middleware("auth");
 
 Route::get("/Auth/Login", function(){
     return view("Auth.Login");
 })->middleware("guest");
+
+Route::post("/Auth/Login/invoke", LoginController::class);
 
 Route::get("/Auth/Register", function(){
     return view("Auth.Register");
@@ -52,7 +55,7 @@ Route::get("/Auth/Register", function(){
 
 Route::post('/Auth/Register/invoke', RegisterController::class);
 
-Route::get("/Auth/LogOut", LogOutController::class)->middleware("auth");
+Route::post("/Auth/LogOut", LogOutController::class)->middleware("auth");
 
 Route::get("/Client/Cart", function(){
     return view("Client.Cart");
@@ -60,7 +63,7 @@ Route::get("/Client/Cart", function(){
 
 Route::get("/Client/Profile", function(){
     return view("Client.Profile");
-});
+})->middleware("auth");
 
 Route::get("/Admin/Dashboard", function(){
     $categoriesCount = count(Category::all());
@@ -70,31 +73,31 @@ Route::get("/Admin/Dashboard", function(){
     $orders = Order::all();
     $ordersCount = count(Order::all());
     return view("Admin.Dashboard", compact("categoriesCount", "productsCount", "rolesCount", "usersCount", "orders", "ordersCount"));
-});
+})->middleware("auth");
 
 
-Route::get("/Admin/Orders", [OrderController::class, "index"])->name("Admin.Orders.Index");
-Route::get("/Admin/Products", [ProductController::class, "index"])->name("Admin.Products.Index");
-Route::get("/Admin/Products/AddProduct", [ProductController::class, "create"]);
-Route::post("/Admin/Products/AddProduct/store", [ProductController::class, "store"]);
-Route::get("/Admin/Products/EditProduct/{product}", [ProductController::class, "edit"]);
-Route::put("/Admin/Products/EditProduct/update/{product}", [ProductController::class, "update"]);
-Route::delete("/Admin/Products/destroy/{product}", [ProductController::class, "destroy"]);
-Route::get("/Admin/Users", [UserController::class, "index"])->name("Admin.Users.Index");
-Route::get("/Admin/Users/AddUser", [UserController::class, "create"]);
-Route::post("Admin/Users/AddUser/store", [UserController::class, "store"]);
-Route::get("/Admin/Users/EditUser/{user}", [UserController::class, "edit"]);
-Route::put("/Admin/Users/EditUser/update/{user}", [UserController::class, "update"]);
-Route::delete("/Admin/Users/destroy/{user}", [UserController::class, "destroy"]);
-Route::get("/Admin/Roles", [RoleController::class, "index"])->name("Admin.Roles.Index");
-Route::get("/Admin/Roles/AddRole", [RoleController::class, "create"]);
-Route::post("/Admin/Roles/AddRole/store", [RoleController::class, "store"]);
-Route::get("/Admin/Roles/EditRole/{role}", [RoleController::class, "edit"]);
-Route::put("/Admin/Roles/EditRole/update/{role}", [RoleController::class, "update"]);
-Route::delete("/Admin/Roles/destroy/{role}", [RoleController::class, "destroy"]);
-Route::get("/Admin/Categories", [CategoryController::class, "index"])->name("Admin.Categories.Index");
-Route::get("/Admin/Categories/AddCategory", [CategoryController::class, "create"]);
-Route::post("/Admin/Categories/AddCategory/store", [CategoryController::class, "store"]);
-Route::get("/Admin/Categories/EditCategory/{category}", [CategoryController::class, "edit"]);
-Route::put("/Admin/Categories/EditCategory/update/{category}", [CategoryController::class, "update"]);
-Route::delete("/Admin/Categories/destroy/{category}", [CategoryController::class, "destroy"]);
+Route::get("/Admin/Orders", [OrderController::class, "index"])->middleware("auth")->name("Admin.Orders.Index");
+Route::get("/Admin/Products", [ProductController::class, "index"])->middleware("auth")->name("Admin.Products.Index");
+Route::get("/Admin/Products/AddProduct", [ProductController::class, "create"])->middleware("auth");
+Route::post("/Admin/Products/AddProduct/store", [ProductController::class, "store"])->middleware("auth");
+Route::get("/Admin/Products/EditProduct/{product}", [ProductController::class, "edit"])->middleware("auth");
+Route::put("/Admin/Products/EditProduct/update/{product}", [ProductController::class, "update"])->middleware("auth");
+Route::delete("/Admin/Products/destroy/{product}", [ProductController::class, "destroy"])->middleware("auth");
+Route::get("/Admin/Users", [UserController::class, "index"])->middleware("auth")->name("Admin.Users.Index");
+Route::get("/Admin/Users/AddUser", [UserController::class, "create"])->middleware("auth");
+Route::post("Admin/Users/AddUser/store", [UserController::class, "store"])->middleware("auth");
+Route::get("/Admin/Users/EditUser/{user}", [UserController::class, "edit"])->middleware("auth");
+Route::put("/Admin/Users/EditUser/update/{user}", [UserController::class, "update"])->middleware("auth");
+Route::delete("/Admin/Users/destroy/{user}", [UserController::class, "destroy"])->middleware("auth");
+Route::get("/Admin/Roles", [RoleController::class, "index"])->middleware("auth")->name("Admin.Roles.Index");
+Route::get("/Admin/Roles/AddRole", [RoleController::class, "create"])->middleware("auth");
+Route::post("/Admin/Roles/AddRole/store", [RoleController::class, "store"])->middleware("auth");
+Route::get("/Admin/Roles/EditRole/{role}", [RoleController::class, "edit"])->middleware("auth");
+Route::put("/Admin/Roles/EditRole/update/{role}", [RoleController::class, "update"])->middleware("auth");
+Route::delete("/Admin/Roles/destroy/{role}", [RoleController::class, "destroy"])->middleware("auth");
+Route::get("/Admin/Categories", [CategoryController::class, "index"])->middleware("auth")->name("Admin.Categories.Index");
+Route::get("/Admin/Categories/AddCategory", [CategoryController::class, "create"])->middleware("auth");
+Route::post("/Admin/Categories/AddCategory/store", [CategoryController::class, "store"])->middleware("auth");
+Route::get("/Admin/Categories/EditCategory/{category}", [CategoryController::class, "edit"])->middleware("auth");
+Route::put("/Admin/Categories/EditCategory/update/{category}", [CategoryController::class, "update"])->middleware("auth");
+Route::delete("/Admin/Categories/destroy/{category}", [CategoryController::class, "destroy"])->middleware("auth");
